backend/transcripter-backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
SQLAlchemy==1.4.27
python-dotenv==0.19.2
Flask-RESTful==0.3.9